delete obsolete code
